+++ .gitignore
@@ -1,2 +1,3 @@
 main.zip
 main
+.vscode
